+++ src/routes/config.js
@@ -3368,7 +3368,7 @@ const Configuration = () => {
                 <Grid container spacing={2} rowSpacing={-0.5}>
                     <Grid item xs={12} md={6}>
                         <Typography variant="body2">
-                            Client: 3.4.0-beta.09 (build.{buildhash})
+                            Client: 3.4.0-beta.10 (build.{buildhash})
                         </Typography>
                     </Grid>
                     <Grid item xs={12} md={6}>